Financial services, payment method, and respect,
Dear, it is a fact that this, Buissnes strategy, I was at the basis of a new announcement, of this international cooperation in the planning of parallel online services, especially with regard to financial services, it is my individual research and planning to  to bring a healthy economy during but especially after covid.  So I have the Feedback where I started the United States to explain an all over Buissnes Strategy International which has been recognized by the world for my protection under code.  When it comes to recognition of rights and freedom, this is not that bad internationally.  But financial services, and even online filing of documents is virtually impossible.  Therefore this detail as a Belgian citizen, I would like to change in this safe way of responsible government policy eg 1, if you want Bank service like becoming a customer, in the first place registration with neighboring country and international as well as european registration are impossible because,  there is a difference between the number of numbers to be filled in as Postal code, or the fact that your citizenship prevents you from entering an Identity, up to the United States.  One has to enter a State such as California otherwise this will not work.  The solution is to start from the country identifiant, you can let the registration form adapt itself, to the rest of such forms.  This also prevents the many incorrect registrations, for example the limit of citizenship when it comes to projects that have been promoted in particular, which prevents them from being able to manage investments that have been invested in international stock exchanges.  This these two adjustments in sequence allows the economy in the international exchanges, down to details such as transportation, to stop reports and misreporting and pull everyone on the same tightrope.  Countering criminal offenders also quickly becomes visible.  That means rapid cooperation on conspicuous, money laundering no longer on a large scale in the beginning, until eventually cooperation international continents together quickly details from their data base information will be surfaced.  These are the two first required objectives to be addressed.  I have a full report financial and buissnes Strategy ready to provide win, win, win result and accessibility financial economics that easily by the right people brings a pleasant work experience.  Also, for example, the differences in rating of investments must be expanded, but for India with Belgium.  Taxes will also allow for better control, and enforcement.  Now that you know that I am the one in humility but have been excellent in financial, strategies, also in covid managed to make profits, and billions of my Project have been invested in the poorest regions because there were no opportunities there.  But because I managed to maintain Christian values, ethics and norms.  There have been reports of groups of specialists fp281460 analyzing online traffic in those circumstances, I knew despite my preparation, like all instruments and tools that change measure each on their scale the same time, then I have to keep priorities in detail  make until after a long time I could make the conclusion how and which Details need at what timing simultaneous attention in priorities because my teachings tell me make yourself comfort Abel in every situation then I have 2 things that everyone has to keep under control that is thinking  and act.  The Rest is Chaos Therefore useless.  Ok I ask for these two parts that are on the first half of new Buissnes plan, and below existing plan of takeover in the deversity and inclusion, what has been my survival in covid is that I managed to build in something completely new strategy with existing resources  .  As my indispensable error margin level had built in.  It keeps setbacks so that you have just in which you know okay in time.  Here's the alarm.  Now my complete individual work was also complete surrender, and drainage of Soul spirit and physical.  All with the wisdom of the corrections contained in the Holy Scriptures, also Art of War, and so on.  If you want extensive news coverage Want health and a win, win, win situation send me Dan and ask the right questions.  Sincerely yours, CEO in senior financial Management business strategy International, politically exposed person, diplomacy inviolable, Ambassador for Human Rights International, Arbitration Director, Licensor, Certificate issuer.  This government now has the lead with these details.  Noel Erik Knaepen Mikael oneal.